+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 Added:
 - Markdown escaping, see [escape example](https://github.com/DrafaKiller/Markdown-dart/blob/main/example/escape.dart)
-- Markdown escape and decoding, using UTF-8 to Hexadecimal, URL-like encoding
+
+Changed:
+- Renamed `placeholder.regex` to `placeholder.pattern`
 
 ## 0.0.2